+++ src/components/shared/CodeViewer.vue
@@ -4,6 +4,8 @@
 
     <div class="code-editor__actions">
       <div class="inline-block">
+        <q-btn round flat dense icon="unfold_more" :title="t('shared.code_editor.actions.expand_all.title')" @click="expandAll" />
+        <q-btn round flat dense icon="unfold_less" :title="t('shared.code_editor.actions.collapse_all.title')" @click="collapseAll" />
         <copy-button round flat dense :custom-handler="copyContent" />
       </div>
 
@@ -36,5 +38,5 @@
   const codeEditorStore = useCodeEditorStore()
 
   const editor: Ref<HTMLElement | null> = ref(null)
-  const { copyContent } = useCodeEditor(editor, { initialValue: toRef(props, 'value') })
+  const { copyContent, collapseAll, expandAll } = useCodeEditor(editor, { initialValue: toRef(props, 'value') })
 </script>