+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * adds update check. this check is also used to anonymously track daily users of elasticvue.
   Check [PRIVACY.md](https://github.com/cars10/elasticvue/blob/master/PRIVACY.md) for details.
 * fix issue when changing search indices - the search query was wrongly reset
+* adds buttons to collapse/expand json, fixes [#291](https://github.com/cars10/elasticvue/issues/291)
 
 ## 1.4.0